Implement new token counting algorithm
We now count backwards, which is quadratic and eliminates the need for
the token estimation hack that uses 4 characters per token. This allows
us to send more context. Unfortunately, we can hit a problem with too
much context, even though the amount is within the 30720 that Google
specifies, so I have limited the context to 29640 tokens.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>