Implement binary search fallback to avoid API request limitations
Loading very long multiturn conversations can still exceed the API
request limit, so we implement a binary search fallback and hope that
there is no limitation on the amount of text we can send.

I am not sure if this is doable via the python library, so we use the
REST API to do it.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>