Refine pruning algorithm
Instead of binary searching by sending a potentially unbounded amount of
context to Google, we use the 4 characters per token approximation to
reduce the context to about 48,000 before binary searching, which limits
the amount of context we send to Google. This is intended to not only
accelerate the binary search, but avoid errors that might occur in the
future if we send excessive context to Google.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>